Move ModifierMatcherFilter to Support namespace
